+++ client/src/components/auth/Signin.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { signin } from "../../api/auth.js";
 import { Link, useNavigate } from "react-router-dom";
-import "../../styles/auth.scss";
 
 const Signin = (): React.ReactElement => {
 	const [checked, setChecked] = useState(false);
@@ -23,60 +22,50 @@ const Signin = (): React.ReactElement => {
 		}
 	};
 	return (
-		<div>
+		<form onSubmit={doSignIn}>
+			<label>
+				<input
+					id="email"
+					type="email"
+					placeholder="email"
+				/>
+			</label>
+			<label>
+				<input
+					id="password"
+					type="password"
+					placeholder="password"
+				/>
+			</label>
+			{	error && 
+				<span>{error}</span>
+			}
+			<button type="submit">
+				Login
+			</button>
+			<label>
+				<input
+					type="checkbox"
+					checked={checked}
+					name="remember"
+					onChange={() => setChecked(!checked)}
+				/>{" "}
+				Remember me
+				<Link
+					to="/forgot"
+					style={{
+						textAlign: "right",
+						display: "inline-block",
+						float: "right",
+					}}
+				>
+					Forgot Password
+				</Link>
+			</label>
 			<div>
-				<div>
-					<Link to="/signin">
-						<button>Sign In</button>
-					</Link>
-					<Link to="/signup">
-						<button>Sign Up</button>
-					</Link>
-				</div>
-				<form onSubmit={doSignIn}>
-					<label>
-						<input
-							id="email"
-							type="email"
-							placeholder="email"
-						/>
-					</label>
-					<label>
-						<input
-							id="password"
-							type="password"
-							placeholder="password"
-						/>
-					</label>
-					{	error && 
-						<span>{error}</span>
-					}
-					<button type="submit">
-						Login
-					</button>
-					<label>
-						<input
-							type="checkbox"
-							checked={checked}
-							name="remember"
-							onChange={() => setChecked(!checked)}
-						/>{" "}
-						Remember me
-						<Link
-							to="/forgot"
-							style={{
-								textAlign: "right",
-								display: "inline-block",
-								float: "right",
-							}}
-						>
-							Forgot Password
-						</Link>
-					</label>
-					{/* <div class="divider">or</div> */}
-				</form>
-			</div>
-		</div>
+				Don't have an account? <Link to="/signup">Sign up</Link>	
+			</div>		
+		</form>
 	);
 }
