+++ client/src/types/AuthModalTypes.tsx
@@ -1,0 +1,8 @@
+interface AuthModalProps {
+    children: React.ReactNode
+}
+
+interface AuthModalState {
+}
+
+export type { AuthModalProps, AuthModalState }
