+++ client/src/types/HomeTypes.tsx
@@ -1,0 +1,8 @@
+interface HomeProps {
+    children: React.ReactNode
+}
+
+interface HomeState {
+}
+
+export type { HomeProps, HomeState }
